capstone-project/pyproject.toml
[project]
name = "capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.8",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "kaggle>=1.8.2",
    "numpy>=2.4.0",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
